+++ dev/Backend/src/main/java/DTO/Records/Image/SolutionValueDTO.java
@@ -1,9 +1,9 @@
 package DTO.Records.Image;
 
+import java.util.List;
+
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 
-import java.util.List;
-
-public record SolutionValueDTO(@NotNull @Valid List<@NotNull String> values, @NotNull int objectiveValue) {
+public record SolutionValueDTO(@NotNull @Valid List<@NotNull String> values, @NotNull double objectiveValue) {
 }
